+++ src/main/java/model/member/BuyerDO.java
@@ -7,7 +7,7 @@ public class BuyerDO {
     private String nickname;
     private String passwd;
     private int point;
-    private int tel;
+    private long tel;
     private String regdate;
     private String buyerImg;
     private String address;
@@ -52,11 +52,11 @@ public class BuyerDO {
         this.point = point;
     }
 
-    public int getTel() {
+    public long getTel() {
         return tel;
     }
 
-    public void setTel(int tel) {
+    public void setTel(long tel) {
         this.tel = tel;
     }
